+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- First iteration of the GUI
+
+Chat tab
+- added delete icon to the last message in the conversation (for easy deletion)
+- added a button in "Advanced Settings" to "Fork a conversation" (save to history for reference, but continue from a fresh copy)
+- added a focus when a template is selected (on expand, on template selection, etc)
+
+Meta-prompting tab
+- Add an experimental meta-prompting experience based on [arxiv](https://arxiv.org/pdf/2401.12954). See the tab "Meta-Prompting" for more details.
 
 ### Fixed
+
+## [0.1.0]
+
+### Added
+- The first iteration of the GUI released